requirements.txt
# Medical RDH — Requirements
# Install: pip install -r requirements.txt

numpy>=1.24.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0
scikit-image>=0.21.0
scipy>=1.11.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Security
pycryptodome>=3.18.0

# Visualization
matplotlib>=3.7.0

# Optional: DICOM support
pydicom>=2.4.0

# Optional: LLM agent
anthropic>=0.25.0

# Optional: experiment tracking
pandas>=2.0.0
